Add RBS helper utilities and extract RBSArg to shared header
